--- src/main/java/models/AmericanGame.java
@@ -1,22 +1,0 @@
-package models;
-
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Random;
-
-/**
- * Created by michaelhilton on 1/25/16.
- */
-public class AmericanGame extends Game {
-
-
-    public void buildDeck() {
-        for(int i = 2; i < 15; i++){
-            deck.add(new Card(i,Suit.Clubs));
-            deck.add(new Card(i,Suit.Hearts));
-            deck.add(new Card(i,Suit.Diamonds));
-            deck.add(new Card(i,Suit.Spades));
-        }
-    }
-
-}
