+++ src/main/java/models/Game.java
@@ -7,7 +7,7 @@ import java.util.Random;
 /*
  * Created by michaelhilton on 1/25/16.
  */
-public abstract class Game {
+public class Game {
 
     public java.util.List<Card> deck = new ArrayList<>();
 
@@ -20,8 +20,14 @@ public abstract class Game {
         this.buildDeck();
     }
 
-
-    public abstract void buildDeck();
+    public void buildDeck() {
+        for(int i = 2; i < 15; i++){
+            deck.add(new Card(i,Suit.Clubs));
+            deck.add(new Card(i,Suit.Hearts));
+            deck.add(new Card(i,Suit.Diamonds));
+            deck.add(new Card(i,Suit.Spades));
+        }
+    }
 
     public void shuffle() {
         long seed = System.nanoTime();
